Simplify FAQ layout
Summary:
The structure faq_layout describes the FAQ and table of contents layout.

This produces the same code, with two single each loop, one for the ToC,
one for the content.

Handlebars seems to require custom partial development for advanced
features like `{{#each <reference>}}`. As a workaround, to keep the
templates as simple as possible, entries from `faq_entries` are
aliased into the `faq_layout` data structure.

To help go back to the table of contents, add buttons to go back up.
The `round` CSS class can be used to create such round bouttons.

This code has received some love so we can offer it as a pattern
to create a FAQ for other static sites.

Test Plan:
Ensure the layout ToC and questions is no-op

Click arrow buttons, ensure the anchor leads us on the top

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3148